Moved template tests from actionpack to actionview
I ensured that all the tests in actionpack still pass, but tests in
actionview fail. The reason for leaving them that way is that I don't
want to move files *and* change them at the same time as it will be
much harder to track changes that way using git log.